Refactor AssignPermissionSetsImpl
- Added function for converting alias to username
- Treat already assigned permsets as successful assignments
- Moved to sfpcommands/permsets
